requirements.txt
Django>=1.9
psycopg2
djangorestframework==3.6.3
djangorestframework-simplejwt==3.2.1
xlrd>=0.9.0
pandas

